exported table to component - click on row to open control
